+++ recipes/setup_clustering.rb
@@ -22,19 +22,6 @@ unless enable_clustering?
   return
 end
 
-# during an initial install, the start/restart commands must deal with accepting
-# the license. So, we must ensure the service[splunk] resource
-# properly deals with the license.
-edit_resource(:service, 'splunk') do
-  action node['init_package'] == 'systemd' ? %i(start enable) : :start
-  supports status: true, restart: true
-  stop_command svc_command('stop')
-  start_command svc_command('start')
-  restart_command svc_command('restart')
-  status_command svc_command('status')
-  provider splunk_service_provider
-end
-
 Chef::Log.debug("Current node clustering mode: #{node['splunk']['clustering']['mode']}")
 
 unless cluster_master?