+++ spec/recipes/setup_clustering_spec.rb
@@ -36,7 +36,7 @@ describe 'chef-splunk::setup_clustering' do
   end
 
   context 'clustering is enabled' do
-    let(:chef_run_init) do
+    let(:runner) do
       ChefSpec::ServerRunner.new do |node|
         node.force_default['dev_mode'] = true
         node.force_default['splunk']['is_server'] = true
@@ -49,40 +49,45 @@ describe 'chef-splunk::setup_clustering' do
 
     context 'invalid cluster mode settings' do
       let(:chef_run) do
-        chef_run_init.node.force_default['splunk']['clustering']['mode'] = 'foo'
-        chef_run_init
+        runner.node.force_default['splunk']['clustering']['mode'] = 'foo'
+        runner.converge(described_recipe) do
+          runner.resource_collection.insert(
+            Chef::Resource::Service.new('splunk', runner.run_context)
+          )
+        end
       end
 
       it 'raises an error' do
-        expect { chef_run.converge(described_recipe) }.to raise_error(RuntimeError, 'Failed to setup clustering: invalid clustering mode')
+        expect { chef_run }.to raise_error(RuntimeError, 'Failed to setup clustering: invalid clustering mode')
       end
     end
 
     context 'cluster master mode' do
       let(:chef_run) do
-        chef_run_init.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
-        chef_run_init.node.run_state['splunk_secret'] = 'notarealsecret'
-        chef_run_init.node.force_default['splunk']['clustering']['mode'] = 'master'
-        chef_run_init
+        runner.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
+        runner.node.run_state['splunk_secret'] = 'notarealsecret'
+        runner.node.force_default['splunk']['clustering']['mode'] = 'master'
+        runner.converge(described_recipe) do
+          runner.resource_collection.insert(
+            Chef::Resource::Service.new('splunk', runner.run_context)
+          )
+        end
       end
 
       context 'default settings (single-site)' do
-        before do
-          chef_run.converge(described_recipe)
-        end
         it_performs 'a successful run', '-mode master -replication_factor 3 -search_factor 2'
       end
 
       context 'multisite clustering with default settings' do
         let(:chef_run) do
-          chef_run_init.node.force_default['splunk']['clustering']['num_sites'] = 2
-          chef_run_init.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
-          chef_run_init.node.run_state['splunk_secret'] = 'notarealsecret'
-          chef_run_init
-        end
-
-        before do
-          chef_run.converge(described_recipe)
+          runner.node.force_default['splunk']['clustering']['num_sites'] = 2
+          runner.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
+          runner.node.run_state['splunk_secret'] = 'notarealsecret'
+          runner.converge(described_recipe) do
+            runner.resource_collection.insert(
+              Chef::Resource::Service.new('splunk', runner.run_context)
+            )
+          end
         end
 
         it_performs 'a successful run', '-mode master -multisite true -available_sites site1,site2 -site site1' \
@@ -91,15 +96,15 @@ describe 'chef-splunk::setup_clustering' do
 
       context 'single-site clustering with custom settings' do
         let(:chef_run) do
-          chef_run_init.node.force_default['splunk']['clustering']['replication_factor'] = 5
-          chef_run_init.node.force_default['splunk']['clustering']['search_factor'] = 3
-          chef_run_init.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
-          chef_run_init.node.run_state['splunk_secret'] = 'notarealsecret'
-          chef_run_init
-        end
-
-        before do
-          chef_run.converge(described_recipe)
+          runner.node.force_default['splunk']['clustering']['replication_factor'] = 5
+          runner.node.force_default['splunk']['clustering']['search_factor'] = 3
+          runner.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
+          runner.node.run_state['splunk_secret'] = 'notarealsecret'
+          runner.converge(described_recipe) do
+            runner.resource_collection.insert(
+              Chef::Resource::Service.new('splunk', runner.run_context)
+            )
+          end
         end
 
         it_performs 'a successful run', '-mode master -replication_factor 5 -search_factor 3'
@@ -107,15 +112,15 @@ describe 'chef-splunk::setup_clustering' do
 
       context 'multisite clustering with custom settings' do
         let(:chef_run) do
-          chef_run_init.node.force_default['splunk']['clustering']['num_sites'] = 3
-          chef_run_init.node.force_default['splunk']['clustering']['site'] = 'site2'
-          chef_run_init.node.force_default['splunk']['clustering']['site_replication_factor'] = 'origin:2,site1:1,site2:1,total:4'
-          chef_run_init.node.force_default['splunk']['clustering']['site_search_factor'] = 'origin:1,site1:1,site2:1,total:3'
-          chef_run_init
-        end
-
-        before do
-          chef_run.converge(described_recipe)
+          runner.node.force_default['splunk']['clustering']['num_sites'] = 3
+          runner.node.force_default['splunk']['clustering']['site'] = 'site2'
+          runner.node.force_default['splunk']['clustering']['site_replication_factor'] = 'origin:2,site1:1,site2:1,total:4'
+          runner.node.force_default['splunk']['clustering']['site_search_factor'] = 'origin:1,site1:1,site2:1,total:3'
+          runner.converge(described_recipe) do
+            runner.resource_collection.insert(
+              Chef::Resource::Service.new('splunk', runner.run_context)
+            )
+          end
         end
 
         it_performs 'a successful run', '-mode master -multisite true -available_sites site1,site2,site3 -site site2' \
@@ -141,20 +146,20 @@ describe 'chef-splunk::setup_clustering' do
       end
 
       before do
-        chef_run_init.node.force_default['splunk']['clustering']['enabled'] = true
-        chef_run_init.node.force_default['splunk']['clustering']['mode'] = 'searchhead'
-        chef_run_init.create_node(cluster_master_node)
+        runner.node.force_default['splunk']['clustering']['enabled'] = true
+        runner.node.force_default['splunk']['clustering']['mode'] = 'searchhead'
+        runner.create_node(cluster_master_node)
       end
 
       context 'default settings (single-site)' do
         let(:chef_run) do
-          chef_run_init.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
-          chef_run_init.node.run_state['splunk_secret'] = 'notarealsecret'
-          chef_run_init
-        end
-
-        before do
-          chef_run.converge(described_recipe)
+          runner.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
+          runner.node.run_state['splunk_secret'] = 'notarealsecret'
+          runner.converge(described_recipe) do
+            runner.resource_collection.insert(
+              Chef::Resource::Service.new('splunk', runner.run_context)
+            )
+          end
         end
 
         it_performs 'a successful run', '-mode searchhead -master_uri https://192.168.0.10:8089 -replication_port 9887'
@@ -162,15 +167,18 @@ describe 'chef-splunk::setup_clustering' do
 
       context 'default settings (multi-site)' do
         let(:chef_run) do
-          chef_run_init.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
-          chef_run_init.node.run_state['splunk_secret'] = 'notarealsecret'
-          chef_run_init.node.force_default['splunk']['clustering']['num_sites'] = 3
-          chef_run_init
+          runner.node.run_state['splunk_auth_info'] = 'admin:notarealpassword'
+          runner.node.run_state['splunk_secret'] = 'notarealsecret'
+          runner.node.force_default['splunk']['clustering']['num_sites'] = 3
+          runner.converge(described_recipe) do
+            runner.resource_collection.insert(
+              Chef::Resource::Service.new('splunk', runner.run_context)
+            )
+          end
         end
 
         before do
           allow_any_instance_of(Chef::Recipe).to receive(:multisite_clustering?).and_return(true)
-          chef_run.converge(described_recipe)
         end
 
         it_performs 'a successful run', '-mode searchhead -site site1 -master_uri https://192.168.0.10:8089 -replication_port 9887'